Xilskey: Type conversion without cast.
This patch fixes below misra-c violation:
        Type conversion without cast.
        MISRA-C:2012 R.10.1,R.10.3,R.10.4,R.11.5


Acked-for-series: VNSL Durga Challa <vnsl.durga.challa@xilinx.com>

+++ lib/sw_services/xilskey/src/xilskey_bbramps_zynqmp.c
@@ -141,7 +141,7 @@ u32 XilSKey_ZynqMp_Bbram_Program(u32 *AesKey)
 
 	if ((StatusRead & (u32)XSK_ZYNQMP_BBRAM_STS_AES_CRC_DONE_MASK)
 							== 0x00U) {
-		Status = XSK_ZYNQMP_BBRAMPS_ERROR_IN_WRITE_CRC;
+		Status = (u32)XSK_ZYNQMP_BBRAMPS_ERROR_IN_WRITE_CRC;
 		goto END;
 	}
